Updated Javadoc Task

+++ Javadoc.command
@@ -1,15 +1,18 @@
-echo cd "$(dirname "$0")"
-cd "$(dirname "$0")"
-echo cd "$(dirname "$0")"/build/libs/FRC-2022
-cd build/libs/FRC-2022
+if [ -z "$(git status --porcelain)" ]; then 
+echo cd FRC-2022
+cd FRC-2022
 echo rm -r frc
 rm -r frc
-echo cp -r ../../../src/main/java/frc frc
-cp -r ../../../src/main/java/frc frc
-echo rm -r ../../../docs
-rm -r ../../../docs
-echo javadoc -subpackages frc -d ../../../docs
-javadoc -subpackages frc -d ../../../docs
-echo "echo theme: jekyll-theme-tactile > ../../../docs/_config.yml" 
-echo theme: jekyll-theme-tactile > ../../../docs/_config.yml
-read -p "Press any key to close" -n 1
+echo cp -r ../src/main/java/frc frc
+cp -r ../src/main/java/frc frc
+echo rm -r ../docs
+rm -r ../docs
+echo javadoc -subpackages frc -d ../docs
+javadoc -subpackages frc -d ../docs
+echo "echo theme: jekyll-theme-tactile > ../docs/_config.yml" 
+echo theme: jekyll-theme-tactile > ../docs/_config.yml
+echo git commit -a -m "Updated Javadoc"
+git commit -a -m "Updated Javadoc"
+else 
+echo Git Directory Not Clean
+fi
